Azure/Functions/ExchangeRateCollector/requirements.txt
azure-functions>=1.18.0
pyodbc>=5.0.0
requests>=2.31.0
python-dotenv>=1.0.0
